+++ components/TwinPaneTab.vue
@@ -67,6 +67,8 @@
 }
 
 .tab-detail {
+  display: flex;
+  flex-direction: column;
   height: calc(100vh - 204px);
 }